feat(Equipe): Migra página da equipe para template e atualiza gatsby-node
